requirements.txt
nltk
pyhyphen